factor uniswap router into helper lib
